🚚(backends) move backend utilities to BackendUtilityMixin
Some methods that were placed in ralph.utils are
data-backend-specific. To allow configuring logging in a
by-backend basis, they accept a logger argument.
Thus, to separate them from ralph's generic utilities and
remove the logger argument from their signature,
we propose to move them to a dedicated BackendUtilityMixin.

We also add default implementations for abstract
`_read_bytes`, `_read_dicts`, `_write_bytes` and `_write_dicts`
methods to promote their reuse and limit repetition.
