+++ src/ralph/backends/data/async_es.py
@@ -16,7 +16,6 @@ from ralph.backends.data.base import (
 )
 from ralph.backends.data.es import ESDataBackend, ESDataBackendSettings, ESQuery
 from ralph.exceptions import BackendException
-from ralph.utils import async_parse_dict_to_bytes, parse_iterable_to_dict
 
 Settings = TypeVar("Settings", bound=ESDataBackendSettings)
 
@@ -155,20 +154,6 @@ class AsyncESDataBackend(
         async for statement in statements:
             yield statement
 
-    async def _read_bytes(
-        self,
-        query: ESQuery,
-        target: Optional[str],
-        chunk_size: int,
-        ignore_errors: bool,
-    ) -> AsyncIterator[bytes]:
-        """Method called by `self.read` yielding bytes. See `self.read`."""
-        statements = self._read_dicts(query, target, chunk_size, ignore_errors)
-        async for statement in async_parse_dict_to_bytes(
-            statements, self.settings.LOCALE_ENCODING, ignore_errors, self.logger
-        ):
-            yield statement
-
     async def _read_dicts(
         self,
         query: ESQuery,
@@ -256,20 +241,6 @@ class AsyncESDataBackend(
         """
         return await super().write(
             data, target, chunk_size, ignore_errors, operation_type, concurrency
-        )
-
-    async def _write_bytes(  # noqa: PLR0913
-        self,
-        data: Iterable[bytes],
-        target: Optional[str],
-        chunk_size: int,
-        ignore_errors: bool,
-        operation_type: BaseOperationType,
-    ) -> int:
-        """Method called by `self.write` writing bytes. See `self.write`."""
-        statements = parse_iterable_to_dict(data, ignore_errors, self.logger)
-        return await self._write_dicts(
-            statements, target, chunk_size, ignore_errors, operation_type
         )
 
     async def _write_dicts(  # noqa: PLR0913
