+++ src/ralph/backends/data/async_mongo.py
@@ -15,7 +15,7 @@ from ralph.backends.data.mongo import (
     MongoQuery,
 )
 from ralph.exceptions import BackendException, BackendParameterException
-from ralph.utils import async_parse_dict_to_bytes, iter_by_batch, parse_iterable_to_dict
+from ralph.utils import iter_by_batch
 
 from ..data.base import (
     AsyncListable,
@@ -164,20 +164,6 @@ class AsyncMongoDataBackend(
         async for statement in statements:
             yield statement
 
-    async def _read_bytes(
-        self,
-        query: MongoQuery,
-        target: Optional[str],
-        chunk_size: int,
-        ignore_errors: bool,
-    ) -> AsyncIterator[bytes]:
-        """Method called by `self.read` yielding bytes. See `self.read`."""
-        statements = self._read_dicts(query, target, chunk_size, ignore_errors)
-        async for statement in async_parse_dict_to_bytes(
-            statements, self.settings.LOCALE_ENCODING, ignore_errors, self.logger
-        ):
-            yield statement
-
     async def _read_dicts(
         self,
         query: MongoQuery,
@@ -234,20 +220,6 @@ class AsyncMongoDataBackend(
         """
         return await super().write(
             data, target, chunk_size, ignore_errors, operation_type, concurrency
-        )
-
-    async def _write_bytes(  # noqa: PLR0913
-        self,
-        data: Iterable[bytes],
-        target: Optional[str],
-        chunk_size: int,
-        ignore_errors: bool,
-        operation_type: BaseOperationType,
-    ) -> int:
-        """Method called by `self.write` writing bytes. See `self.write`."""
-        statements = parse_iterable_to_dict(data, ignore_errors, self.logger)
-        return await self._write_dicts(
-            statements, target, chunk_size, ignore_errors, operation_type
         )
 
     async def _write_dicts(  # noqa: PLR0913
